Refactored PrisonerEventProcessor to PrisonerEventManager
Replaced PrisonerEventProcessor with PrisonerEventManager for better clarity and organization. Introduced new enums (ResponseQuality, EventResultEffect, PrisonerEvent) to improve code structure and event handling. Adjusted affected classes and properties to align with the new logic.
