+++ MekHQ/src/mekhq/campaign/personnel/prisoners/enums/EventResultEffect.java
@@ -1,0 +1,5 @@
+package mekhq.campaign.personnel.prisoners.enums;
+
+public enum EventResultEffect {
+    NONE, UNIQUE, QUIRK, INJURY_MINOR, INJURY_MAJOR, DEATH_GUARD, DEATH_PRISONER, NEW_SKILL
+}
