+++ MekHQ/src/mekhq/campaign/personnel/prisoners/enums/PrisonerEvent.java
@@ -1,0 +1,11 @@
+package mekhq.campaign.personnel.prisoners.enums;
+
+public enum PrisonerEvent {
+    ARGUMENT, WILD_STORIES, TAMPERING, CONVERSATIONS, RATIONS, TRADE, DRAINED, RESCUE, REPAIRS,
+    SICKNESS, VETERAN, GRAFFITI, PRAYER, BARTERING, OFFICER, DICE, LOVERS, SOBBING, PROPAGANDA,
+    SONGS, REFUSE_RATIONS, PLOTTING, EQUIPMENT, PLANNED_RESCUE, MISTAKE, LETTER, ILLNESS,
+    DO_ANDROIDS_DREAM_OF_ELECTRONIC_SHEEP, HEAVY_METAL, PARANOIA, TUNNEL, PET_RODENT, ESCAPE_ROPE,
+    UNDERCOVER, BRAND, SILENCE, SINGING, PAPER, HOLIDAY, WHISPERS, LEADER, ARGUMENTS, SENTIMENTAL_ITEM,
+    MALNUTRITION, INJURIES, TERROR, SCREAMING, PHOTO, GHOSTS, VOICES, BREAKOUT, RIOT, MURDER, FIRE,
+    POISON, HOSTAGE, BOMB, EXECUTION, ABANDONED_TO_DIE, UNITED
+}
