+++ MekHQ/src/mekhq/campaign/personnel/prisoners/enums/ResponseQuality.java
@@ -1,0 +1,5 @@
+package mekhq.campaign.personnel.prisoners.enums;
+
+public enum ResponseQuality {
+    RESPONSE_NEUTRAL, RESPONSE_POSITIVE, RESPONSE_NEGATIVE
+}
